requirements.txt
altgraph==0.17.4
asgiref==3.9.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.2.5
djangorestframework==3.16.1
environs==14.3.0
greenlet==3.2.3
h11==0.14.0
idna==3.6
Imagegrab==0.0.3
iniconfig==2.0.0
marshmallow==4.0.0
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==23.2
pefile==2023.2.7
pillow==10.4.0
playwright==1.54.0
pluggy==1.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pyee==13.0.0
PyGetWindow==0.0.9
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
pytest==8.0.0
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dotenv==1.1.0
python-slugify==8.0.4
pytweening==1.2.0
pywin32-ctypes==0.2.2
requests==2.32.3
selenium==4.17.2
setuptools==69.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.5.3
text-unidecode==1.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2025.2
urllib3==2.2.0
webdriver-manager==4.0.2
wsproto==1.2.0
